+++ threedi_models_and_simulations/data_models/simulation_data_models.py
@@ -108,6 +108,11 @@ class Laterals(SimulationElement):
 
 
 @dataclass
+class Substances(SimulationElement):
+    data: list = None
+
+
+@dataclass
 class DWF(SimulationElement):
     data: dict = None
 
@@ -199,6 +204,7 @@ class NewSimulation:
     structure_controls: StructureControls = None
     initial_conditions: InitialConditions = None
     laterals: Laterals = None
+    substances: Substances = None
     dwf: DWF = None
     breach: Breach = None
     precipitation: Precipitation = None